+++ Changelog.md
@@ -1,5 +1,5 @@
 ## 0.0.50
-*INCOMPLETE*
++ Fixed import errors in `fxn.compile` module.
 
 ## 0.0.49
 + Fixed import errors in `fxn.beta` module.